@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -26,12 +26,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
 
-  flutter_svg: ^0.19.3
-  get: ^3.26.0
-  firebase_dynamic_links: ^0.8.0
-  url_launcher: ^6.0.3
+  flutter_svg: ^0.22.0
+  get: ^4.3.8
+  firebase_dynamic_links: ^2.0.9
+  url_launcher: ^6.0.12
 dev_dependencies:
   flutter_test:
     sdk: flutter
