@@ -26,10 +26,10 @@ class WebMenu extends StatelessWidget {
 
 class WebMenuItem extends StatefulWidget {
   const WebMenuItem({
-    Key key,
-    @required this.isActive,
-    @required this.text,
-    @required this.press,
+    Key? key,
+    required this.isActive,
+    required this.text,
+    required this.press,
   }) : super(key: key);
 
   final bool isActive;
